omics_graph_learning/requirements.txt
cmapPy==4.0.1
cmapPy==4.0.1
cooler==0.9.3
cooltools==0.6.1
joblib==1.4.2
matplotlib==3.5.2
networkx==2.8.4
optuna==3.5.0
plotly==5.19.0
pybedtools==0.9.0
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.4.2
scipy==1.13.0
seaborn==0.11.2
torch_geometric==2.5.3
tqdm==4.66.4
